Re-factored code to separate presentation from object walking
